detectZombie2でテキストが出力されない問題を修正．併せて戻り値をlong型に修正．

+++ control.h
@@ -39,7 +39,7 @@ void *exeDetectMobs();
 void killPython(void);
 
 int detectZombie1(void);
-int detectZombie2(void);
+long detectZombie2(void);
 int detectZombie3(void);
 int detectMobsArray(int mode , int ibuf[]);
 int detectMobs(int mode);
